pr: make ktlint and detekt happy
